jiff_core::civil

Module date

Structs

Date

A Gregorian civil date.

ISOWeekDate

An ISO 8601 civil week date.

UnixEpochDay

A civil date represented by a number of days since the Unix epoch (1970-01-01).

Functions

iso_week_start_from_year 🔒

Returns the Unix epoch day corresponding to the first day in the ISO 8601 week year given.