jiff_core

Module tz

Building blocks for supporting time zones.

Modules

posix

Implements POSIX time zone string parsing and transition handling.

tzif

Implements TZif time zone parsing and transition handling.

Structs

AmbiguousError

An error that occurs when an unmabiguous civil datetime is demanded.

AmbiguousTimestamp

A possibly ambiguous Timestamp.

Offset

A fixed offset, in seconds, from UTC.

OffsetInfo

Information associated with an offset when doing a time zone transition lookup.

Transition

A representation a single time zone transition.

Enums

AmbiguousOffset

A possibly ambiguous Offset.

Dst

An enum indicating whether a particular datetime is in daylight saving time (DST) or not.

Functions

offset

Creates a new time zone offset in a const context from a given number of hours.

Type Aliases

Abbreviation

A type that defines the storage for a time zone abbreviation.

TimeZoneId

A type that defines the storage for an IANA time zone identifier.