jiff_core::tz::posix

Struct Parser

struct Parser<'s> {
    tz: &'s [u8],
    pos: Cell<usize>,
    ianav3plus: bool,
}

A parser for POSIX time zones.

Fields
tz: &'s [u8]

The TZ string that we’re parsing.

pos: Cell<usize>

The parser’s current position in tz.

ianav3plus: bool

Whether to use IANA rules, i.e., when parsing a TZ string in a TZif file of version 3 or greater. From tzfile(5):

First, the hours part of its transition times may be signed and range from -167 through 167 instead of the POSIX-required unsigned values from 0 through 24. Second, DST is in effect all year if it starts January 1 at 00:00 and ends December 31 at 24:00 plus the difference between daylight saving and standard time.

At time of writing, I don’t think I understand the significance of the second part above. (RFC 8536 elaborates that it is meant to be an explicit clarification of something that POSIX itself implies.) But the first part is clear: it permits the hours to be a bigger range.

Implementations

impl<'s> Parser<'s>

fn new<B: ?Sized + AsRef<[u8]>>(tz: &'s B) -> Parser<'s>

Create a new parser for extracting a POSIX time zone from the given bytes.

fn parse(&self) -> Result<TimeZone, ParseError>

Parses a POSIX time zone from the current position of the parser and ensures that the entire TZ string corresponds to a single valid POSIX time zone.

fn parse_prefix(&self) -> Result<(TimeZone, usize), ParseError>

Parses a POSIX time zone from the current position of the parser and returns the remaining input.

fn parse_posix_time_zone(&self) -> Result<TimeZone, ParseError>

Parse a POSIX time zone from the current position of the parser.

Upon success, the parser will be positioned immediately following the TZ string.

fn parse_posix_dst(&self, std_offset: Offset) -> Result<Dst, ParseError>

Parse a DST zone with an optional explicit transition rule.

This assumes the parser is positioned at the first byte of the DST abbreviation.

Upon success, the parser will be positioned immediately after the end of the DST transition rule (which might just be the abbreviation, but might also include explicit start/end datetime specifications).

fn parse_abbreviation(&self) -> Result<Abbreviation, AbbreviationError>

Parse a time zone abbreviation.

This assumes the parser is positioned at the first byte of the abbreviation. This is either the first character in the abbreviation, or the opening quote of a quoted abbreviation.

Upon success, the parser will be positioned immediately following the abbreviation name.

The string returned is guaranteed to be no more than 30 bytes. (This restriction is somewhat arbitrary, but it’s so we can put the abbreviation in a fixed capacity array.)

fn parse_unquoted_abbreviation( &self, ) -> Result<Abbreviation, UnquotedAbbreviationError>

Parses an unquoted time zone abbreviation.

This assumes the parser is position at the first byte in the abbreviation.

Upon success, the parser will be positioned immediately after the last byte in the abbreviation.

The string returned is guaranteed to be no more than 30 bytes. (This restriction is somewhat arbitrary, but it’s so we can put the abbreviation in a fixed capacity array.)

fn parse_quoted_abbreviation( &self, ) -> Result<Abbreviation, QuotedAbbreviationError>

Parses a quoted time zone abbreviation.

This assumes the parser is positioned immediately after the opening < quote. That is, at the first byte in the abbreviation.

Upon success, the parser will be positioned immediately after the closing > quote.

The string returned is guaranteed to be no more than 30 bytes. (This restriction is somewhat arbitrary, but it’s so we can put the abbreviation in a fixed capacity array.)

fn parse_posix_offset(&self) -> Result<Offset, OffsetError>

Parse a POSIX time offset.

This assumes the parser is positioned at the first byte of the offset. This can either be a digit (for a positive offset) or the sign of the offset (which must be either - or +).

Upon success, the parser will be positioned immediately after the end of the offset.

fn parse_rule(&self) -> Result<Rule, RuleError>

Parses a POSIX DST transition rule.

This assumes the parser is positioned at the first byte in the rule. That is, it comes immediately after the DST abbreviation or its optional offset.

Upon success, the parser will be positioned immediately after the DST transition rule. In typical cases, this corresponds to the end of the TZ string.

fn parse_posix_datetime(&self) -> Result<DayTime, DateTimeError>

Parses a POSIX datetime specification.

This assumes the parser is position at the first byte where a datetime specification is expected to occur.

Upon success, the parser will be positioned after the datetime specification. This will either be immediately after the date, or if it’s present, the time part of the specification.

fn parse_posix_date(&self) -> Result<Day, DateError>

Parses a POSIX date specification.

This assumes the parser is positioned at the first byte of the date specification. This can be J (for one based Julian day without leap days), M (for “weekday of month”) or a digit starting the zero based Julian day with leap days. This routine will validate that the position points to one of these possible values. That is, the caller doesn’t need to parse the M or the J or the leading digit. The caller should just call this routine when it expect a date specification to follow.

Upon success, the parser will be positioned immediately after the date specification.

fn parse_posix_julian_day_no_leap(&self) -> Result<i16, JulianNoLeapError>

Parses a POSIX Julian day that does not include leap days (1 <= n <= 365).

This assumes the parser is positioned just after the J and at the first digit of the Julian day. Upon success, the parser will be positioned immediately following the day number.

fn parse_posix_julian_day_with_leap(&self) -> Result<i16, JulianLeapError>

Parses a POSIX Julian day that includes leap days (0 <= n <= 365).

This assumes the parser is positioned at the first digit of the Julian day. Upon success, the parser will be positioned immediately following the day number.

fn parse_weekday_of_month( &self, ) -> Result<(i8, i8, Weekday), WeekdayOfMonthError>

Parses a POSIX “weekday of month” specification.

This assumes the parser is positioned just after the M byte and at the first digit of the month. Upon success, the parser will be positioned immediately following the “weekday of the month” that was parsed.

The tuple returned is month (1..=12), week (1..=5) and weekday (0..=6 with 0=Sunday).

fn parse_posix_time(&self) -> Result<TransitionCivilTime, TimeError>

This parses a POSIX time specification in the format [+/-]hh?[:mm[:ss]].

This assumes the parser is positioned at the first h (or the sign, if present). Upon success, the parser will be positioned immediately following the end of the time specification.

fn parse_month(&self) -> Result<i8, MonthError>

Parses a month.

This is expected to be positioned at the first digit. Upon success, the parser will be positioned after the month (which may contain two digits).

fn parse_week(&self) -> Result<i8, WeekOfMonthError>

Parses a week-of-month number.

This is expected to be positioned at the first digit. Upon success, the parser will be positioned after the week digit.

fn parse_weekday(&self) -> Result<Weekday, WeekdayError>

Parses a weekday number.

This is expected to be positioned at the first digit. Upon success, the parser will be positioned after the week digit.

The weekday returned is guaranteed to be in the range 0..=6, with 0 corresponding to Sunday.

fn parse_hour_ianav3plus(&self) -> Result<i16, HourIanaError>

Parses an hour from a POSIX time specification with the IANA v3+ extension. That is, the hour may be in the range 0..=167. (Callers should parse an optional sign preceding the hour digits when IANA V3+ parsing is enabled.)

The hour is allowed to be a single digit (unlike minutes or seconds).

This assumes the parser is positioned at the position where the first hour digit should occur. Upon success, the parser will be positioned immediately after the last hour digit.

fn parse_hour_posix(&self) -> Result<i8, HourPosixError>

Parses an hour from a POSIX time specification, with the allowed range being 0..=24.

The hour is allowed to be a single digit (unlike minutes or seconds).

This assumes the parser is positioned at the position where the first hour digit should occur. Upon success, the parser will be positioned immediately after the last hour digit.

fn parse_minute(&self) -> Result<i8, MinuteError>

Parses a minute from a POSIX time specification.

The minute must be exactly two digits.

This assumes the parser is positioned at the position where the first minute digit should occur. Upon success, the parser will be positioned immediately after the second minute digit.

fn parse_second(&self) -> Result<i8, SecondError>

Parses a second from a POSIX time specification.

The second must be exactly two digits.

This assumes the parser is positioned at the position where the first second digit should occur. Upon success, the parser will be positioned immediately after the second second digit.

fn parse_number_with_exactly_n_digits( &self, n: usize, ) -> Result<i32, NumberError>

Parses a signed 64-bit integer expressed in exactly n digits.

If n digits could not be found (or if the TZ string ends before n digits could be found), then this returns an error.

This assumes that n >= 1 and that the parser is positioned at the first digit. Upon success, the parser is positioned immediately after the nth digit.

fn parse_number_with_upto_n_digits(&self, n: usize) -> Result<i32, NumberError>

Parses a signed 64-bit integer expressed with up to n digits and at least 1 digit.

This assumes that n >= 1 and that the parser is positioned at the first digit. Upon success, the parser is position immediately after the last digit (which can be at most n).

fn parse_optional_sign(&self) -> Result<Option<i8>, OptionalSignError>

Parses an optional sign.

This assumes the parser is positioned at the position where a positive or negative sign is permitted. If one exists, then it is consumed and returned. Moreover, if one exists, then this guarantees that it is not the last byte in the input. That is, upon success, it is valid to call self.byte().

impl<'s> Parser<'s>

Helper routines for parsing a POSIX TZ string.

fn bump(&self) -> bool

Bump the parser to the next byte.

If the end of the input has been reached, then false is returned.

fn is_done(&self) -> bool

Returns true if the next call to bump would return false.

fn byte(&self) -> u8

Return the byte at the current position of the parser.

This panics if the parser is positioned at the end of the TZ string.

fn maybe_byte(&self) -> Option<u8>

Return the byte at the current position of the parser. If the TZ string has been exhausted, then this returns None.

fn pos(&self) -> usize

Return the current byte offset of the parser.

The offset starts at 0 from the beginning of the TZ string.