jiff_core::tz

Struct Offset

pub struct Offset { /* private fields */ }

A fixed offset, in seconds, from UTC.

Implementations

impl Offset

pub const MIN: Offset

The minimum possible offset from UTC.

pub const MAX: Offset

The maximum possible offset from UTC.

pub const UTC: Offset

The UTC offset.

pub const ZERO: Offset

The zero offset.

pub const fn constant(hours: i8) -> Offset

Creates a new time zone offset in a const context from a given number of hours.

pub const fn constant_seconds(seconds: i32) -> Offset

Creates a new time zone offset in a const context from a given number of seconds.

pub const fn from_hours(hours: i8) -> Result<Offset, RangeError>

Creates a new time zone offset from a given number of hours.

Negative offsets correspond to time zones west of the prime meridian, while positive offsets correspond to time zones east of the prime meridian. Equivalently, in all cases, civil-time - offset = UTC.

pub const fn from_seconds(seconds: i32) -> Result<Offset, RangeError>

Returns a new time zone offset from UTC given its representation in seconds.

An error is also returned when seconds is not in the range specified by OffsetTotalSeconds.

pub const fn seconds(self) -> i32

Returns the seconds value corresponding to this time zone offset.

pub const fn negate(self) -> Offset

Returns the negation of this offset.

A negative offset will become positive and vice versa. This is a no-op if the offset is zero.

This never panics.

pub const fn signum(self) -> i8

Returns the “sign number” or “signum” of this offset.

The number returned is -1 when this offset is negative, 0 when this offset is zero and 1 when this span is positive.

pub const fn is_positive(self) -> bool

Returns true if and only if this offset is positive.

This returns false when the offset is zero or negative.

pub const fn is_negative(self) -> bool

Returns true if and only if this offset is less than zero.

This returns false when the offset is zero or positive.

pub const fn is_zero(self) -> bool

Returns true if and only if this offset is zero.

Or equivalently, when this offset corresponds to Offset::UTC.

pub const fn checked_add(self, seconds: i32) -> Result<Offset, RangeError>

Adds the given number of seconds to this offset.

If the resulting offset would be outside the an offset’s boundaries, an error is returned.

pub const fn checked_sub(self, seconds: i32) -> Result<Offset, RangeError>

Subtracts the given number of seconds from this offset.

If the resulting offset would be outside the an offset’s boundaries, an error is returned.

pub const fn until(self, other: Offset) -> i32

Returns the number of seconds from this offset to other.

pub const fn since(self, other: Offset) -> i32

Returns the number of seconds since this offset from other.

pub const fn to_datetime(self, timestamp: Timestamp) -> DateTime

Converts a Unix timestamp with an offset to a Gregorian datetime.

The offset should correspond to the number of seconds required to add to this timestamp to get the local time.

pub const fn to_timestamp(self, dt: DateTime) -> Result<Timestamp, RangeError>

Converts the given civil datetime to a timestamp using this offset.

Errors

This returns an error if this would have returned a timestamp outside of its minimum and maximum values.