jiff_core::tz

Type Alias Abbreviation

pub type Abbreviation = SmallStr<TIME_ZONE_ABBREVIATION_MAX>;

A type that defines the storage for a time zone abbreviation.

For an abbreviation whose length is less than or equal to a certain implementation defined number, it will be stored inline inside an array. All other cases spill out into the heap. (Unless callers do not have dynamic memory allocation, in which case, whatever tried to store the time zone abbreviation will return an error. For example, parsing a POSIX time zone transition rule will fail in that case.)

Aliased Type

pub struct Abbreviation { /* private fields */ }