jiff_core::tz

Module tzif

Implements TZif time zone parsing and transition handling.

The TZif format is used by the Time Zone Database.

The parser in this module is designed to handle untrusted input. That is, there is no input that should cause it to panic or allocate memory in a way that is not proportional to the size of the TZif data.

These binary files are the ones commonly found in Unix distributions in the /usr/share/zoneinfo directory.

Structs

DateTime

The representation for a civil datetime used by the TZif implementation.

LocalTimeType

A “local time type” from TZif data.

MaybeNamedTimeZone

A representation of a possibly named time zone backed by TZif data.

ParseError

An error that can occur when parsing TZif data.

TimeZone

A representation of an unnamed time zone backed by TZif data.

Timestamp

The representation for a timestamp used by the TZif implementation.

TransitionInfo

TZif transition info beyond the timestamp and civil datetime.

Transitions

The set of transitions in TZif data, laid out in column orientation.

Enums

Indicator

The possible indicator values for standard/wall and UT/local.

TransitionKind

The kind of a transition.

Functions

is_possibly_tzif

Returns true if the data might be in TZif format.