jiff_core::tz::tzif::parser

Struct ParsedTimeZone

struct ParsedTimeZone {
    fixed: ParsedFixed,
    types: Vec<LocalTimeType>,
    transitions: ParsedTransitions,
}

Fields
fixed: ParsedFixedtypes: Vec<LocalTimeType>transitions: ParsedTransitions
Implementations

impl ParsedTimeZone

fn parse(bytes: &[u8]) -> Result<TimeZone, ParseError>

fn new(version: u8) -> ParsedTimeZone

fn finish(self) -> Result<TimeZone, ParseError>

fn parse32<'b>( header32: Header, bytes: &'b [u8], ) -> Result<(ParsedTimeZone, &'b [u8]), ParseError>

fn parse64<'b>( header32: Header, bytes: &'b [u8], ) -> Result<(ParsedTimeZone, &'b [u8]), ParseError>

fn parse_transitions<'b>( &mut self, header: &Header, bytes: &'b [u8], ) -> Result<&'b [u8], ParseError>

fn parse_transition_types<'b>( &mut self, header: &Header, bytes: &'b [u8], ) -> Result<&'b [u8], TransitionTypeError>

fn parse_local_time_types<'b>( &mut self, header: &Header, bytes: &'b [u8], ) -> Result<&'b [u8], ParseError>

fn parse_time_zone_designations<'b>( &mut self, header: &Header, bytes: &'b [u8], ) -> Result<&'b [u8], TimeZoneDesignatorError>

fn parse_leap_seconds<'b>( &mut self, header: &Header, bytes: &'b [u8], ) -> Result<&'b [u8], ParseError>

Parses leap second corrections in TZif data.

Jiff ignores leap seconds, so this only validates that the byte ranges described by the header are well-formed.

fn parse_indicators<'b>( &mut self, header: &Header, bytes: &'b [u8], ) -> Result<&'b [u8], IndicatorError>

fn parse_footer<'b>( &mut self, _header: &Header, bytes: &'b [u8], ) -> Result<&'b [u8], FooterError>

fn verify_posix_time_zone_consistency( &self, ) -> Result<(), InconsistentPosixTimeZoneError>

Validates that the POSIX TZ string we parsed, if present, is consistent with the last transition in this time zone.

fn add_civil_datetimes_to_transitions(&mut self)

Add civil datetimes to our transitions.

This speeds up time zone lookups when the input is a civil datetime.

fn fatten(&mut self)

Fatten up this TZif data with additional transitions from the POSIX time zone footer.

fn fatten_add_transition( &mut self, posix_tz: &TimeZone, prev: Timestamp, ) -> Option<Timestamp>

fn find_or_create_local_time_type( &mut self, offset: Offset, abbrev: &str, dst: Dst, ) -> Option<u8>

fn find_or_create_designation(&mut self, needle: &str) -> Option<u8>

Looks for a designation matching needle and returns its offset.

If one is not found, then a new designation is added and its new offset is returned.

If the offset of the designation exceeds u8::MAX, then this returns None. This is not possible for designations in a RFC 9636 TZif file, since their offsets are always a single byte big. However, it is theoretically possible for this to return None on conforming data when fattening is enabled. In which case, it’s possible for new local time types to be added beyond the maximum. But this would require its POSIX time zone to use a local time type not found in the file already.

fn designation(&self, typ: &LocalTimeType) -> &str